requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pyFirmata2==2.4.5
pyserial==3.5
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
